// src/public/js/types/personalinformationitemstype.ts
/**
 * HTML tag items (label and values of profile personal information section)
 */
export type PersonalInformationItemsType = {
    name_label_el: JQuery<HTMLElement>;
    name_value_el: JQuery<HTMLElement>;
    surname_label_el: JQuery<HTMLElement>;
    surname_value_el: JQuery<HTMLElement>;
    sex_label_el: JQuery<HTMLElement>;
    sex_value_el: JQuery<HTMLElement>;
    birth_date_label_el: JQuery<HTMLElement>;
    birth_date_value_el: JQuery<HTMLElement>;
    birth_place_label_el: JQuery<HTMLElement>;
    birth_place_value_el: JQuery<HTMLElement>;
    living_place_label_el: JQuery<HTMLElement>;
    living_place_value_el: JQuery<HTMLElement>;
    secondary_school_label_el: JQuery<HTMLElement>;
    secondary_school_value_el: JQuery<HTMLElement>;
    university_label_el: JQuery<HTMLElement>;
    university_value_el: JQuery<HTMLElement>;
    work_label_el: JQuery<HTMLElement>;
    work_value_el: JQuery<HTMLElement>;
    telephone_label_el: JQuery<HTMLElement>;
    telephone_value_el: JQuery<HTMLElement>;
    email_label_el: JQuery<HTMLElement>;
    email_value_el: JQuery<HTMLElement>;
}